feat: support dataclasses with 'field()' definitions
The @gql.type annotation internally wraps classes to be dataclasses.
Unfortunately, this happens whether or not the original class is already
a dataclass.

A small example of how this fails:

```python
from dataclasses import dataclass, field

from dacite import from_dict
from strawberry_django_plus import gql

@gql.type
@dataclass
class Vehicle:
    powertrain: str
    pcm_modules: list[str] = field(
        default_factory=list,
    )

data = {
    "powertrain": "hybrid",
}

print(from_dict(Vehicle, data))
```

This patch adds an optional parameter that will skip wrapping a class as
a dataclass. I did originally try checking to see if the `cls` object
was a dataclass (with `dataclasses.is_dataclass()`) but this caused
issues with some built-in Strawberry types. I also tried copying over
the original field definitions or `default_factory`/`default`
attributes, but I wasn't able to find a way to access this information
on the original dataclass. A way to detect this case and handle it would
certainly be preferred to adding yet another flag.